Adding ComponentMenu
