+++ ui/app/src/modules/Preview/ToolBar.tsx
@@ -36,22 +36,11 @@ import { useDispatch } from 'react-redux';
 import { changeSite } from '../../state/actions/sites';
 import { Site } from '../../models/Site';
 import { LookupTable } from '../../models/LookupTable';
-import {
-  useActiveSiteId,
-  useEnv,
-  usePreviewGuest,
-  usePreviewState,
-  useSelection,
-  useSpreadState
-} from '../../utils/hooks';
+import { useActiveSiteId, useEnv, usePreviewState, useSelection } from '../../utils/hooks';
 import { getHostToGuestBus } from './previewContext';
-import { isBlank, popPiece } from '../../utils/string';
+import { isBlank } from '../../utils/string';
 import { FormattedMessage } from 'react-intl';
-import { Menu } from '@material-ui/core';
-import { palette } from '../../styles/theme';
-import EmbeddedLegacyEditors from './EmbeddedLegacyEditors';
-import { getItem } from '../../services/content';
-import PublishDialog from '../Content/Publish/PublishDialog';
+import ComponentMenu from '../../components/ComponentMenu';
 
 const foo = () => void 0;
 
@@ -99,10 +88,6 @@ const useStyles = makeStyles((theme: Theme) => createStyles({
     display: 'flex',
     alignItems: 'center'
   },
-  separator : {
-    borderTop: `1px solid ${palette.gray.light3}`,
-    borderBottom: `1px solid ${palette.gray.light3}`
-  },
   iframe: {
     height: '0',
     border: 0,
@@ -146,37 +131,6 @@ export function AddressBar(props: AddressBarProps) {
   const noSiteSet = isBlank(site);
   const [internalUrl, setInternalUrl] = useState(url);
   const [anchorEl, setAnchorEl] = useState(null);
-  const guest = usePreviewGuest();
-  const contentTypesBranch = useSelection(state => state.contentTypes);
-  const contentTypes = contentTypesBranch.byId ? Object.values(contentTypesBranch.byId) : null;
-  const AUTHORING_BASE = useSelection<string>(state => state.env.AUTHORING_BASE);
-  const defaultSrc = `${AUTHORING_BASE}/legacy/form?`;
-  const [dialogConfig, setDialogConfig] = useSpreadState({
-    open: false,
-    src: null,
-    type: null,
-    inProgress: true
-  });
-
-  const [publishDialog, setPublishDialog] = useSpreadState({
-    open: false,
-    item: null,
-    scheduling: null
-  });
-
-
-  useEffect(() => {
-    if (guest && guest.models) {
-      getItem(site, guest.models[guest.modelId].craftercms.path).subscribe(
-        (item) => {
-          setPublishDialog({ item })
-        },
-        (error) => {
-          console.log(error)
-        }
-      );
-    }
-  }, [guest, setPublishDialog, site]);
 
   useEffect(() => {
     (url) && setInternalUrl(url);
@@ -186,58 +140,6 @@ export function AddressBar(props: AddressBarProps) {
 
   const handleClick = (event: any) => {
     setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const getPath = (type?: string) => {
-    switch (type) {
-      case 'publish':
-      case 'form': {
-        return guest.models[guest.modelId].craftercms.path;
-      }
-      case 'template': {
-        return contentTypes.find((contentType) => contentType.id === guest.models[guest.modelId].craftercms.contentType).displayTemplate;
-      }
-      case 'controller': {
-        let pageName = popPiece(guest.models[guest.modelId].craftercms.contentType, '/');
-        return `/scripts/pages/${pageName}.groovy`;
-      }
-      default: {
-        return guest.models[guest.modelId].craftercms.path;
-      }
-    }
-  };
-
-  const handleEdit = (type: string) => {
-    handleClose();
-    switch (type) {
-      case 'schedule': {
-        setPublishDialog({ open: true, scheduling: 'custom' });
-        break;
-      }
-      case 'publish': {
-        setPublishDialog({ open: true, scheduling: 'now' });
-        break;
-      }
-      case 'form':
-      case 'template':
-      case 'controller': {
-        setDialogConfig(
-          {
-            open: true,
-            src: `${defaultSrc}site=${site}&path=${getPath(type)}&type=${type}`,
-            type
-          });
-        break;
-      }
-    }
-  };
-
-  const onClosePublish = (response) => {
-    setPublishDialog({ open: false, scheduling: null, item:{...publishDialog.item, isLive: true} });
   };
 
   return (
@@ -290,80 +192,11 @@ export function AddressBar(props: AddressBarProps) {
       <IconButton className={classes.iconButton} aria-label="search" onClick={handleClick}>
         <MoreVertRounded/>
       </IconButton>
-      <Menu
+      <ComponentMenu
         anchorEl={anchorEl}
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
-      >
-        <MenuItem onClick={() => handleEdit('form')}>
-          <FormattedMessage
-            id="previewToolBar.menu.edit"
-            defaultMessage="Edit"
-          />
-        </MenuItem>
-        {
-          (!publishDialog.item?.lockOwner && !publishDialog.item?.isLive) &&
-          <MenuItem onClick={() => handleEdit('schedule')}>
-            <FormattedMessage
-              id="previewToolBar.menu.schedule"
-              defaultMessage="Schedule"
-            />
-          </MenuItem>
-        }
-        {
-          (!publishDialog.item?.lockOwner && !publishDialog.item?.isLive) &&
-          <MenuItem onClick={() => handleEdit('publish')}>
-            <FormattedMessage
-              id="previewToolBar.menu.publish"
-              defaultMessage="Publish"
-            />
-          </MenuItem>
-        }
-        <MenuItem onClick={handleClose}>
-          <FormattedMessage
-            id="previewToolBar.menu.history"
-            defaultMessage="History"
-          />
-        </MenuItem>
-        <MenuItem onClick={handleClose}>
-          <FormattedMessage
-            id="previewToolBar.menu.dependencies"
-            defaultMessage="Dependencies"
-          />
-        </MenuItem>
-        <MenuItem onClick={handleClose}>
-          <FormattedMessage
-            id="previewToolBar.menu.delete"
-            defaultMessage="Delete"
-          />
-        </MenuItem>
-        <MenuItem onClick={handleClose} className={classes.separator}>
-          <FormattedMessage
-            id="previewToolBar.menu.infoSheet"
-            defaultMessage="Info Sheet"
-          />
-        </MenuItem>
-        <MenuItem onClick={() => handleEdit('template')}>
-          <FormattedMessage
-            id="previewToolBar.menu.editTemplate"
-            defaultMessage="Edit Template"
-          />
-        </MenuItem>
-        <MenuItem onClick={() => handleEdit('controller')}>
-          <FormattedMessage
-            id="previewToolBar.menu.editController"
-            defaultMessage="Edit Controller"
-          />
-        </MenuItem>
-      </Menu>
-      {
-        dialogConfig.open &&
-        <EmbeddedLegacyEditors dialogConfig={dialogConfig} setDialogConfig={setDialogConfig} getPath={getPath}/>
-      }
-      {
-        publishDialog.open &&
-        <PublishDialog scheduling={publishDialog.scheduling} items={[publishDialog.item]} onClose={onClosePublish}/>
-      }
+        setAnchorEl={setAnchorEl}
+        site={site}
+      />
     </>
   );
 }
